[AutowireArrayParameter] Decouple new package
