+++ composer.json
@@ -13,6 +13,7 @@
         "j7mbo/twitter-api-php": "^1.0",
         "jean85/pretty-package-versions": "^1.2",
         "latte/latte": "^2.5",
+        "nette/application": "^3.0",
         "nette/di": "^3.0",
         "nette/finder": "^2.5",
         "nette/neon": "^3.0",
@@ -21,15 +22,16 @@
         "ocramius/package-versions": "^1.4",
         "phar-io/version": "^2.0.1",
         "phpstan/phpdoc-parser": "^0.4",
+        "phpstan/phpstan": "^0.12",
         "psr/simple-cache": "^1.0",
         "slevomat/coding-standard": "^6.0",
         "squizlabs/php_codesniffer": "^3.5",
         "symfony/cache": "^4.4|^5.0",
         "symfony/config": "^4.4|^5.0",
         "symfony/console": "^4.4|^5.0",
-        "symfony/debug": "^4.4|^5.0",
         "symfony/dependency-injection": "^4.4|^5.0",
         "symfony/event-dispatcher": "^4.4|^5.0",
+        "symfony/filesystem": "^4.3|^5.0",
         "symfony/finder": "^4.4|^5.0",
         "symfony/http-kernel": "^4.4|^5.0",
         "symfony/process": "^4.4|^5.0",
@@ -38,24 +40,23 @@
         "twig/twig": "^3.0"
     },
     "require-dev": {
+        "doctrine/cache": "dev-master",
         "doctrine/doctrine-bundle": "^2.0",
+        "doctrine/doctrine-cache-bundle": "^1.3",
         "doctrine/orm": "^2.7",
         "doctrine/persistence": "^1.3.3",
-        "doctrine/cache": "dev-master",
-        "doctrine/doctrine-cache-bundle": "^1.3",
-        "nette/application": "^3.0",
         "phpstan/phpstan": "^0.12",
         "phpunit/phpunit": "^8.4",
         "symfony/framework-bundle": "^4.4|^5.0",
         "symfony/templating": "^4.4|^5.0",
         "symfony/translation": "^4.4|^5.0",
         "symfony/twig-bundle": "^4.4|^5.0",
-        "tracy/tracy": "^2.7",
-        "rector/rector": "dev-master"
+        "tracy/tracy": "^2.7"
     },
     "autoload": {
         "psr-4": {
             "Symplify\\Autodiscovery\\": "packages/Autodiscovery/src",
+            "Symplify\\AutowireArrayParameter\\": "packages/AutowireArrayParameter/src",
             "Symplify\\ChangelogLinker\\": "packages/ChangelogLinker/src",
             "Symplify\\CodingStandard\\": "packages/CodingStandard/src",
             "Symplify\\CodingStandard\\TokenRunner\\": "packages/CodingStandard/packages/TokenRunner/src",
@@ -94,6 +95,7 @@
         ],
         "psr-4": {
             "Symplify\\Autodiscovery\\Tests\\": "packages/Autodiscovery/tests",
+            "Symplify\\AutowireArrayParameter\\Tests\\": "packages/AutowireArrayParameter/tests",
             "Symplify\\ChangelogLinker\\Tests\\": "packages/ChangelogLinker/tests",
             "Symplify\\CodingStandard\\Tests\\": "packages/CodingStandard/tests",
             "Symplify\\CodingStandard\\TokenRunner\\Tests\\": "packages/CodingStandard/packages/TokenRunner/tests",
@@ -109,8 +111,8 @@
             "Symplify\\NeonToYamlConverter\\Tests\\": "packages/NeonToYamlConverter/tests",
             "Symplify\\PHPStanExtensions\\Tests\\": "packages/PHPStanExtensions/tests",
             "Symplify\\PackageBuilder\\Tests\\": "packages/PackageBuilder/tests",
-            "Symplify\\SmartFileSystem\\Tests\\": "packages/SmartFileSystem/tests",
             "Symplify\\SetConfigResolver\\Tests\\": "packages/SetConfigResolver/tests",
+            "Symplify\\SmartFileSystem\\Tests\\": "packages/SmartFileSystem/tests",
             "Symplify\\Statie\\Generator\\Tests\\": "packages/Statie/packages/Generator/tests",
             "Symplify\\Statie\\HeadlineAnchorLinker\\Tests\\": "packages/Statie/packages/HeadlineAnchorLinker/tests",
             "Symplify\\Statie\\MigratorJekyll\\Tests\\": "packages/Statie/packages/MigratorJekyll/tests",
@@ -122,21 +124,22 @@
         }
     },
     "replace": {
-        "symplify/autodiscovery": "7.1",
+        "symplify/autodiscovery": "self.version",
         "symplify/better-phpdoc-parser": "7.1",
-        "symplify/changelog-linker": "7.1",
-        "symplify/coding-standard": "7.1",
-        "symplify/easy-coding-standard": "7.1",
-        "symplify/easy-coding-standard-tester": "7.1",
-        "symplify/flex-loader": "7.1",
-        "symplify/latte-to-twig-converter": "7.1",
-        "symplify/monorepo-builder": "7.1",
-        "symplify/neon-to-yaml-converter": "7.1",
-        "symplify/package-builder": "7.1",
-        "symplify/phpstan-extensions": "7.1",
-        "symplify/smart-file-system": "7.1",
-        "symplify/set-config-resolver": "7.1",
-        "symplify/statie": "7.1"
+        "symplify/changelog-linker": "self.version",
+        "symplify/coding-standard": "self.version",
+        "symplify/easy-coding-standard": "self.version",
+        "symplify/easy-coding-standard-tester": "self.version",
+        "symplify/flex-loader": "self.version",
+        "symplify/latte-to-twig-converter": "self.version",
+        "symplify/monorepo-builder": "self.version",
+        "symplify/neon-to-yaml-converter": "self.version",
+        "symplify/package-builder": "self.version",
+        "symplify/phpstan-extensions": "self.version",
+        "symplify/smart-file-system": "self.version",
+        "symplify/set-config-resolver": "self.version",
+        "symplify/statie": "self.version",
+        "symplify/autowire-array-parameter": "self.version"
     },
     "scripts": {
         "complete-check": [
@@ -162,6 +165,6 @@
     "config": {
         "sort-packages": true
     },
-    "prefer-stable": true,
-    "minimum-stability": "dev"
+    "minimum-stability": "dev",
+    "prefer-stable": true
 }
