+++ packages/AutowireArrayParameter/.gitattributes
