+++ packages/AutowireArrayParameter/.github/FUNDING.yml
@@ -1,0 +1,3 @@
+# These are supported funding model platforms
+
+patreon: rectorphp
