+++ packages/AutowireArrayParameter/composer.json
@@ -1,0 +1,32 @@
+{
+    "name": "symplify/autowire-array-parameter",
+    "description": "Autowire array parameters in for your Symfony applications",
+    "license": "MIT",
+    "require": {
+        "php": "^7.2",
+        "nette/application": "^3.0",
+        "nette/utils": "^3.0",
+        "symfony/dependency-injection": "^4.4|^5.0",
+        "symplify/package-builder": "^7.2"
+    },
+    "require-dev": {
+        "phpunit/phpunit": "^8.4"
+    },
+    "autoload": {
+        "psr-4": {
+            "Symplify\\AutowireArrayParameter\\": "src"
+        }
+    },
+    "autoload-dev": {
+        "psr-4": {
+            "Symplify\\AutowireArrayParameter\\Tests\\": "tests"
+        }
+    },
+    "extra": {
+        "branch-alias": {
+            "dev-master": "7.2-dev"
+        }
+    },
+    "minimum-stability": "dev",
+    "prefer-stable": true
+}
