+++ packages/AutowireArrayParameter/tests/DependencyInjection/CompilerPass/AutowireArrayParameterCompilerPassTest.php
@@ -1,0 +1,19 @@
+<?php declare(strict_types=1);
+
+namespace Symplify\AutowireArrayParameter\Tests\DependencyInjection\CompilerPass;
+
+use Symplify\AutowireArrayParameter\Tests\HttpKernel\AutowireArrayParameterHttpKernel;
+use Symplify\AutowireArrayParameter\Tests\Source\SomeCollector;
+use Symplify\PackageBuilder\Tests\AbstractKernelTestCase;
+
+final class AutowireArrayParameterCompilerPassTest extends AbstractKernelTestCase
+{
+    public function test(): void
+    {
+        $this->bootKernel(AutowireArrayParameterHttpKernel::class);
+
+        /** @var SomeCollector $collector */
+        $collector = self::$container->get(SomeCollector::class);
+        $this->assertCount(2, $collector->getCollected());
+    }
+}
