+++ packages/AutowireArrayParameter/tests/Source/Collected/FirstCollected.php
@@ -1,0 +1,10 @@
+<?php declare(strict_types=1);
+
+namespace Symplify\AutowireArrayParameter\Tests\Source\Collected;
+
+use Symplify\AutowireArrayParameter\Tests\Source\Contract\CollectedInterface;
+
+final class FirstCollected implements CollectedInterface
+{
+
+}
