+++ packages/AutowireArrayParameter/tests/Source/Contract/CollectedInterface.php
@@ -1,0 +1,8 @@
+<?php declare(strict_types=1);
+
+namespace Symplify\AutowireArrayParameter\Tests\Source\Contract;
+
+interface CollectedInterface
+{
+
+}
