+++ packages/AutowireArrayParameter/upgrade/rector/autowire-array-parameter-72.yaml
@@ -1,0 +1,3 @@
+services:
+    Rector\Renaming\Rector\Class_\RenameClassRector:
+        Symplify\PackageBuilder\DependencyInjection\CompilerPass\AutowireArrayParameterCompilerPass: 'Symplify\AutowireArrayParameter\DependencyInjection\CompilerPass\AutowireArrayParameterCompilerPass'
